requirements.txt
pipecat-ai
python-dotenv
aiohttp
websockets
flask
twilio
